% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{est.pair}
\alias{est.pair}
\title{Identify pairwise local maxima and local minima of the second-order derivative}
\usage{
est.pair(vall, peak, gamma)
}
\arguments{
\item{vall}{vector of locations of significant local minima}

\item{peak}{vector of locations of significant local maxima}

\item{gamma}{bandwidth of Gaussian kernel smoothing function}
}
\value{
a list of detected pairs and detected change-point locations through second-order derivative testing
}
\description{
Identify pairwise local maxima and local minima of the second-order derivative
}
