% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{est.slope}
\alias{est.slope}
\title{Estimate piecewise slope for piecewise linear signal}
\usage{
est.slope(x, breaks)
}
\arguments{
\item{x}{vector of raw data}

\item{breaks}{vector of change-point locations}
}
\value{
a vector of piecewise slope
}
\description{
Estimate piecewise slope for piecewise linear signal
}
