% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Fdr}
\alias{Fdr}
\title{Compute TPR and FPR}
\usage{
Fdr(uh, th, b)
}
\arguments{
\item{uh}{numerical vector of estimated change point locations}

\item{th}{numerical vector of true change point locations}

\item{b}{location tolerance, usually specified as the bandwidth \code{gamma}}
}
\value{
a dataframe of \code{FDR} (FPR) and \code{Power} (TPR)
}
\description{
Compute TPR and FPR
}
