% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrBH.R
\name{fdrBH}
\alias{fdrBH}
\title{FDR threshold based on the Benjamini-Hochberg algorithm}
\usage{
fdrBH(p, q)
}
\arguments{
\item{p}{a vector of p-values}

\item{q}{False Discovery Rate level}
}
\value{
p-value threshold based on independence or positive dependence
}
\description{
FDR threshold based on the Benjamini-Hochberg algorithm
}
\examples{
fdrBH(seq(0.01,0.1,0.01),q=0.1)
}
