% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HST_stock}
\alias{HST_stock}
\title{Stock price of Host & Hotel Resorts (HST)}
\format{
A data frame with 2,517 rows and 6 columns:
\describe{
  \item{Date}{date from November 7, 2011 to November 5, 2021}
  \item{Close, Open, High, Low}{stock price}
  \item{Volume}{stock exchange volume}
}
}
\source{
<https://finance.yahoo.com/quote/HST>
}
\usage{
HST_stock
}
\description{
A subset of daily stock price data of HST from November 7, 2011 to November
5, 2021.
}
\keyword{datasets}
