% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{snr}
\alias{snr}
\title{Compute SNR of a certain location in signal-plus-noise data}
\usage{
snr(order, gamma, is.jump, jump, diffb, addb)
}
\arguments{
\item{order}{order of derivative of data}

\item{gamma}{bandwidth of Gaussian kernel}

\item{is.jump}{logical value indicating if the location to be tested is a jump}

\item{jump}{jump value}

\item{diffb}{difference of the slopes at two sides of the location}

\item{addb}{sum of the slopes, only used when order is 1}
}
\value{
a scalar of SNR
}
\description{
Compute SNR of a certain location in signal-plus-noise data
}
