% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichPeaks.R
\name{which.peaks}
\alias{which.peaks}
\title{Find local maxima or minima}
\usage{
which.peaks(x, partial = FALSE, decreasing = FALSE)
}
\arguments{
\item{x}{a vector with maxima or minima}

\item{partial}{if 'true' endpoints will be considered}

\item{decreasing}{find local minima if 'true', ortherwise local maxima}
}
\value{
a vector of positions of local maxima or minima
}
\description{
Find local maxima or minima
}
\examples{
a = 100:1
which.peaks(a*sin(a/3))
}
