% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{est.slope}
\alias{est.slope}
\title{Estimate piecewise slope for piecewise linear model}
\usage{
est.slope(x, breaks)
}
\arguments{
\item{x}{numerical vector of signal-plus-noise data}

\item{breaks}{numerical vector of change-point locations}
}
\value{
a vector of estimated piecewise slope
}
\description{
Estimate piecewise slope for piecewise linear model
}
