% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smth.gau}
\alias{smth.gau}
\title{Smoothing data using Gaussian kernel}
\usage{
smth.gau(x, gamma)
}
\arguments{
\item{x}{numeric vector of values to smooth}

\item{gamma}{bandwidth of Gaussian kernel}
}
\value{
vector of smoothed values
}
\description{
Smoothing data using Gaussian kernel
}
\examples{
smth.gau(x=rnorm(1000), gamma=20)
}
