% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{snr}
\alias{snr}
\title{Compute SNR of a certain change point location}
\usage{
snr(order, gamma, is.jump, jump, diffb, addb)
}
\arguments{
\item{order}{order of derivative of data}

\item{gamma}{bandwidth of Gaussian kernel}

\item{is.jump}{logical value indicating if the location to be calculated is a jump point}

\item{jump}{jump height}

\item{diffb}{difference of the slopes on left and right sides of the location}

\item{addb}{sum of the slopes, only used when order is 1}
}
\value{
value of SNR
}
\description{
Compute SNR of a certain change point location
}
