% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichPeaks.R
\name{which.peaks}
\alias{which.peaks}
\title{Find local maxima and local minima of data sequence}
\usage{
which.peaks(x, partial = FALSE, decreasing = FALSE)
}
\arguments{
\item{x}{numerical vector contains local maxima (minima)}

\item{partial}{logical value indicating if the two endpoints will be considered}

\item{decreasing}{logical value indicating whether to find local minima}
}
\value{
a vector of locations of local maxima or minima
}
\description{
Find local maxima and local minima of data sequence
}
\examples{
a = 100:1
which.peaks(a*sin(a/3))
}
